* build: use more modern form for AM_INIT_AUTOMAKE
With recent automake versions, the macro AM_INIT_AUTOMAKE only takes a
list of flags as first argument. The version and project name should be
specified in AC_INIT and this was already the case. Moreover, we specify
we are not a GNU project (so we don't require some mandatory files).