* build: ensure "make dist" works
A tarball can be created with "make dist" and should be equivalent to
what was done until now. This can be checked with "make distcheck". The
change is quite invasive since automake has to know each file.

To avoid repetition, we use libtool convenience libraries. Moreover,
plugins dependencies were specified directly in configure.ac. Instead,
we use automake conditionals. Same for the threadpool.

We also need to rename crc32.c to crc32.h since it's not possible to
compile it as a standalone unit.