* uacctd.c: switch support from ULOG to NFLOG
NFLOG supports both IPv4 and IPv6. While ULOG is still supported in
recent kernels, NFLOG is supported since 2.6.14 and there is little
point to support both.

Each occurrence of "ulog" has been replaced by "nflog". However,
"uacctd" is still called this way and configuration options remain the
same (but the semantics change a bit for some of them). This should
minimize the disruption for downstreams and users (but users will still
have to update the netfilter rules to use NFLOG instead of ULOG).

To test, one can use the following iptables setup:

    iptables -t mangle -I POSTROUTING -j NFLOG --nflog-group 4
    iptables -t raw -I PREROUTING -j NFLOG --nflog-group 4

And use the following command:

    ./src/uacctd -c in_iface,out_iface,src_mac,dst_mac,src_host,dst_host,proto,src_port,dst_port -P print -g 4