.travis.yml: do not run hub_docker on PR
For security reasons, stage hub_docker should never be attempted
to run on a PR. Fortunately, travis is already wise enough not
to load the credentials to not screw things badly, but it makes
the pipeline fail - always - for PRs.

Make sure this stage is only run on legit commits.