Autogenerate pmacct-build.h
Maintaining, by hand, `pmacct-build.h` is a waste of Paolo's
precious time. This commit autogenerates - in the same format
- the `pmacct-build.h` build string, by using BUILT_SOURCES and
a shell script Makefile recipe.

For compilations which don't have .git/, it's assumed to be a
RELEASE, so BUILD=="RELEASE".

I would recommend to change the format of `pmacct-build.h` to the
git commit hash instead.