external_libs: build deps sys, add libcdada
This commit does two major things:

1) Adds external_libs/ folder, and prepares the build system to
   accept building, as part of pmacct's `make` process, external
   dependencies. These external dependencies are in the form of
   git submodules, which are cloned and checkedout as part of
   the build process.
2) Adds libcdada submodule as a dependency of pmacct, and also
   as a showcase of the build system, to later be used by the
   code.

A bit more details on 2). `src/external_libs/Makefile.am` contains
the basics that do the "work" of cloning the subrepos, checking out
the commits and compiling the dependencies. These dependencies are
compiled before any pmacct object is attempted to be compiled.

Dependencies _shall_ be compiled and linked by pmacct daemons
statically, as they are NOT distributed as part of pmacct binaries,
just statically linked.

Dependencies will only be attempted to be compiled if the git
submodule commit hash changes, or it hasn't been built before. To
do so, a "mark" file is created under $(builddir)/src/external_libs/.

The dependency build system puts the results under
$(builddir)/src/external_libs/rootfs, with the typical structure
of:

* bin/
* include/
* lib/

When compiled in-tree, this directory, and the compilation marks
are git ignored.

If pmacct is compiled from a tarball (no git repository), dependency
building is skipped. There will be more refinement on this later on