configure.ac: add --without-external-deps option
This commit adds --without-external-deps (and --with) to control
whether to build the git submodule dependencies (libraries or not).

This option is particularly useful for OS package maintainers,
specially if existing dependencies (libavro, kafka...) are
integrated in the build system as submodules.

When --without-external-deps is used, `-lcdada` is required to
be installed in the system.