external_libs: fix bug on submodule commit update
Fix a bug in which if you pulled new commits from pmacct and
submodules (libcdada) changed, the new version of the submodule
was not compiled.

This was due to an incorrect check against the checkout submodule
and not the one being pointed by the supermodule (pmacct) commit.

To prevent `make` to sync the commit - only for developers, the
env special env. variable:

```
export PMACCT_EXT_LIBS_DONT_SYNC=1
```

Can be set, and will skip `git submodule update`