ci: move to Github Actions
travis-ci.org is dead, and travis.com is also dead, as they are
putting all the possible problems for OSS - even though they claim
they still fully support OSS. Sigh...

Time to move on. This commit replicates the _same_ exact CI
structure that we had in travis-ci.org into Github Actions (GHA).

It also removes all travis related files.

For this PR to work, the following "secrets" have to be defined in
Github settings of the project:

* `DOCKER_USERNAME`
* `DOCKER_PASSWORD`

Which are the same ones already defined in travis-ci.org and are the
username (pmacct) and the token for dockerhub uploads.

As a final remark, this commit adds nightly builds at 12am, everyday
in the default branch.