Add a markdown-ified version of the quickstart guide.
I have converted the actual text into Markdown for more readability.

The work is not completed, as there are parts that still need some
refactoring in Markdown, but it's better to have the file in the repo as
soon as possible so new changes might get into this one rather than on
the text file (which I kept separate for preserving history).