ci: split docker and dockerhub upload jobs
Prior to this commit both the generation of docker images and
uploading them to dockerhub was done in the same job. This meant
that docker generation was skipped for all PRs.

In order to solve this, this commit splits this in two sequential
jobs, one to compile the docker images (and exporting the registry
as artifact) and one publishing them to dockerhub (having imported
the registry artifact). The former will also run for PRs.