ci: generate dockers in parallel to build-and-test
Prior to this commit, docker generation was happening _after_
build-and-test stage.

With this commit pmacct-docker (pmacct docker container generation)
happens in parallel to builders+build-and-test, and only upload
awaits _both_ build-and-test and pmacct-docker, as we don't want
to publish (bleeding-edge tag) images that don't pass all tests.

This cuts some minutes in the pipeline.

Thanks @dcaba for the suggestion