pmacct: use corret IP option for jumbo payload packets
According to RFC 2292 'Advanced Sockets API for IPv6' datagrams
with more than 65535 bytes payload require the jumbo payload option,
e.g the length of the payload ist stored in the hop-by-hop header.
In order to exclude jumbo payload packets we have to check if the
IP6 header payload length is zero and the next header is a
hop-by-hop header.