kafka: move schema registration from writer to plugin process
The schema structure is defined by the program configuration
and does not change during runtime. Registering the same schema
information on every start of a writer process does not give any
advantage and creates a high load on the schema registry.
To reduce the load on the registry the schema registration is
moved to the plugin process and done only once at program start.