log: avoid to cut off log messages in a logfile
When log messages are longer than our message buffer the message
gets cut off. As the trailing newline also gets cut off the
message will be concatenated with the following message which
makes the log hard to read. Avoid to cut off the messages by not
using an intermediate buffer.