ci: add Datahangar E2E test
This commit adds a nightly run of pmacct `HEAD` of `master` running
a full e2e data pipeline (based on the https://github.com/datahangar/datahangar
project, which runs natively in K8s), to better mimic a real scenario.
This hopefully automates the work that I've been doing manually for
the last months :).

The current coverage is _very_ minimal, but sufficient to add _some_
value already.

It tests the following features of `nfacctd`:

* Netflow ingestion. Fields actively tested:
  * "timestamp_start", "timestamp_end", "etype", "src_as", "in_iface",
    "peer_src_ip", "peer_dst_ip", "peer_src_as", "peer_dst_as", "as_path",
     "dst_as", "src_net", "src_mask", "dst_net", "dst_mask", "src_host",
     "src_host_country", "dst_host", "dst_host_country"
* Output to Kafka using - at the moment - SCRAM auth.
* Multiple instances of the Kafka plugin.
* Deployability in K8s environments

The immediate next steps are:

* Being more rigorous with counters (Paolo&I are chasing a potential
  bug in accounting as a result of this test, see:
  https://github.com/datahangar/datahangar/blob/main/.github/actions/test/files/count-db-records.yaml#L24).
* Add BGP and test flow enrichment using BGP.
* Use Kafka with encryption
* Test geolocation plugin.