dh_e2h: fix use bleeding-edge instead of SHA
Images are not tagged with the SHA, only bleeding-edge, latest and
vX.Y.Z. Therefore, and since this workflow is only run nightly,
run with `bleeding-edge`.

It might be that in the future we want to also tag with the SHA to
E2E test after freezing a version in both branches (but that would
require changes in this workflow too, to run on other events).