Fix inconsistency of command line options
uacctd's -N was removed in 445f96cd but still present in the optstring.

pmacct's -t was defined in the optstring with an required argument, but
it takes none. The argument was wrongly "inherited" in 45c860a776 by the
removal of -m.

pmacct's -P was missing in the usage output since the beginning in
1f7b559c. Because it's unused since f78fe098, remove the option
completely.

pmtelemetryd's -Z was removed in e3cce045 but still present in the usage
output.